% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\docType{methods}
\name{combineData,CNAclinicData,CNAclinicData-method}
\alias{combineData,CNAclinicData,CNAclinicData-method}
\title{Combine the slots of two CNAclinicData objects}
\usage{
\S4method{combineData}{CNAclinicData,CNAclinicData}(x, y)
}
\arguments{
\item{x}{a \code{\link{CNAclinicData}} object.}

\item{y}{a \code{\link{CNAclinicData}} object.}
}
\value{
Returns an object of class \code{\link{CNAclinicData}} with
sample data of both \code{x} and \code{y}.
}
\description{
This method uses cbind to create a new CNAcliniData object that contains the
sample data of both arguments x and y. Both CNAcliniData objects to be
combined (x, y)  must have the same collection of slots filled and have
identical bin annotations.
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
