% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{exportData}
\alias{exportData}
\title{Export data}
\usage{
exportData(x, dataType, geneInfo = NULL, fileName = NULL,
  fileType = c("csv", "igv"), filterNA = TRUE, digits = 3, ...)
}
\arguments{
\item{x}{X}

\item{dataType}{X}

\item{geneInfo}{X}

\item{fileName}{X}

\item{fileType}{X}

\item{filterNA}{X}

\item{digits}{X}

\item{...}{X}
}
\description{
Export data
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
\author{
Dineika Chandrananda
}
