% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Get transcript and position information for NCBI Entrez gene IDs}
\usage{
getGeneInfo(geneID, genome, transcript = "first")
}
\arguments{
\item{geneID}{a \code{\link[base]{numeric}} vector containing NCBI Entrez
gene IDs}

\item{genome}{Genome build used to align sequencing reads.}

\item{select}{When select is \code{"first"} (default), the first
transcript of the input genes are chosen. If \code{"all"}, information of
all possible transcripts will be contained in output.}
}
\value{
a data.frame containing the information for the \code{geneID} input.
The data.frame has the columns named \code{chromosome}, \code{start}, \code{end}
}
\description{
A helper function to generate gene information as input for plotSampleData
function.
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
