% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimalBinsize.R
\name{optimalBinsize}
\alias{optimalBinsize}
\title{Assess optimal genomic bin size to partition read counts.}
\usage{
optimalBinsize(bamfiles = NULL, bamnames = NULL, pathToBams = NULL,
  binSizes = c(10, 30, 50, 100, 250, 500, 750, 1000), measure = "CV",
  lineColor = "red4", chromosomesFilter = c("X", "Y", "M", "MT"),
  savePlot = FALSE, plotPrefix = "optimalBinsize", minMapq = 20,
  isPaired = NA, isProperPair = NA, isUnmappedQuery = FALSE,
  hasUnmappedMate = NA, isMinusStrand = NA, isMateMinusStrand = NA,
  isFirstMateRead = NA, isSecondMateRead = NA, isSecondaryAlignment = NA,
  isDuplicate = FALSE)
}
\arguments{
\item{bamfiles}{A \code{\link[base]{character}} vector of BAM file names
with or without full path. If NULL (default), all files with extension
.bam, are read from directory path.}

\item{bamnames}{An optional \code{\link[base]{character}} vector of sample
names. Defaults to file names with extension \code{.bam} removed.}

\item{pathToBams}{If \code{bamfiles} is NULL, all files ending with ".bam"
extension will be read from this path.}

\item{binSizes}{A \code{\link[base]{numeric}} vector of genomic bin sizes,
in units of kilo base pairs (1000 base pairs),
e.g. \code{binSizes = c(10, 30, 50)} corresponds to bins of 10, 30 and 50 kbp
bins.}

\item{measure}{The goodness of fit criteria (AIC or CV). Defaults to "CV".}

\item{lineColor}{Line color to use in plot.}

\item{chromosomesFilter}{A \code{\link[base]{character}} vector specifying
which chromosomes to filter out.
Defaults to the sex chromosomes and mitochondrial reads,
i.e. \code{c("X", "Y", "M", "MT")}. Use NA to use all chromosomes.}

\item{savePlot}{if TRUE (default) saves plots of each sample to
working directory.}

\item{plotPrefix}{Prefix for plot title and pdf file name. Defaults to "optimalBinsize".}

\item{minMapq}{If quality scores exists, the minimum quality score required
in order to keep a read (20, default).}

\item{isPaired}{A \code{\link[base]{logical}}(1) indicating whether unpaired
(FALSE), paired (TRUE), or any (NA, default) read should be returned.}

\item{isProperPair}{A \code{\link[base]{logical}}(1) indicating whether
improperly paired (FALSE), properly paired (TRUE), or any (NA, default) read
should be returned.}

\item{isUnmappedQuery}{A \code{\link[base]{logical}}(1) indicating whether
unmapped (TRUE), mapped (FALSE, default), or any (NA) read should be returned.}

\item{hasUnmappedMate}{A \code{\link[base]{logical}}(1) indicating whether
reads with mapped (FALSE), unmapped (TRUE), or any (NA, default) mate should
be returned.}

\item{isMinusStrand}{A \code{\link[base]{logical}}(1) indicating whether
reads aligned to the plus (FALSE), minus (TRUE), or any (NA, default) strand
should be returned.}

\item{isMateMinusStrand}{A \code{\link[base]{logical}}(1) indicating whether
mate reads aligned to the plus (FALSE), minus (TRUE), or any (NA, default)
strand should be returned.}

\item{isFirstMateRead}{A \code{\link[base]{logical}}(1) indicating whether
the first mate read should be returned (TRUE) or not (FALSE), or whether mate
read number should be ignored (NA, default).}

\item{isSecondMateRead}{A \code{\link[base]{logical}}(1) indicating whether
the second mate read should be returned (TRUE) or not (FALSE), or whether
mate read number should be ignored (NA, default).}

\item{isSecondaryAlignment}{A \code{\link[base]{logical}}(1) indicating
whether alignments that are primary (FALSE), are not primary (TRUE) or whose
primary status does not matter (NA, default) should be returned.}

\item{isDuplicate}{A \code{\link[base]{logical}}(1) indicating that
un-duplicated (FALSE, default), duplicated (TRUE), or any (NA) reads should
be returned.}
}
\value{
Returns a list. The first element is a data.frame holding information of the
average read counts per bin size, the other elements are sample-specific
\code{\link[ggplot2]{ggplot}} objects.
}
\description{
Calculate Akaike's information criterion (AIC) and cross-validation (CV)
log-likelihood to infer the optimal bin size to partition read counts across
genome.
}
\details{
As a guidance, choose bin sizes which have low AIC and/or high CV values
but also contain 30-180 read counts on average. This strikes a reasonable
balance between error variability and bias of CNA. Using a much smaller bin
size may result in many genomic regions with zero read count and make the
overall analysis non-informative. At the other extreme, using a much bigger
bin size will 'smooth out' some pattern of alteration (i.e. increasing bias).
The process of estimating the optimal bin size is in the context of
low-coverage sequence data, so use sensible values for the binSizes argument
when the input data is not of shallow whole-genome depth (<10 million reads).
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
\seealso{
Internally, the function \code{\link[NGSoptwin]{opt.win.onesample}}
of the \pkg{NGSoptwin} package is used.
}
\author{
Dineika Chandrananda
}
