% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSampleData.R
\docType{methods}
\name{plotMultiSampleData,CNAclinicData-method}
\alias{plotMultiSampleData,CNAclinicData-method}
\title{Plots heat map of a cohort of samples}
\usage{
\S4method{plotMultiSampleData}{CNAclinicData}(object, geneInfo = NULL,
  clusterGenes = FALSE, clusterSamples = FALSE, dataType = "summary",
  chromosomesFilter = c("X", "Y", "M", "MT"), outlierLog2R = 6,
  showCalls = TRUE, callThreshLog2R = c(-0.15, 0.15), callLegend = FALSE,
  legendSize = 12, legendKeySize = 2, legendPosition = "bottom",
  legendHoriz = TRUE, lossColor = "orange2", gainColor = "blue",
  pointColor = "black", neutralColor = "white", missingColor = "grey20",
  xlab = ifelse(is.null(geneInfo), "Chromosome", "Gene"), xaxt = "s",
  xaxSize = 10, ylab = NULL, yaxSize = 10, main = NULL, mainSize = 10)
}
\arguments{
\item{object}{CNAclinicData.}
}
\value{
a ggplot2 object
}
\description{
Plots heat map of a cohort of samples
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
