% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleData.R
\docType{methods}
\name{plotSampleData,CNAclinicData-method}
\alias{plotSampleData,CNAclinicData-method}
\title{plotSampleData}
\usage{
\S4method{plotSampleData}{CNAclinicData}(object, geneInfo = NULL,
  clusterGenes = FALSE, chromosomesFilter = c("X", "Y", "M", "MT"),
  outlierLog2R = 5, showLog2RPlot = TRUE, showHeatmap = FALSE,
  showDataPoints = TRUE, showSegments = TRUE, segmentType = c("summary"),
  segHMMColor = "green4", segCBSColor = "green", segPLSColor = "purple",
  segLACBSColor = "hotpink", segSummaryColor = "red2",
  segmentLineWidth = rep(1, length(segmentType)), segmentLineType = rep(1,
  length(segmentType)), segmentLegend = showSegments, legendSize = 12,
  legendKeySize = 2, segmentLegendPosition = "bottom",
  segmentLegendHoriz = TRUE, showCalls = TRUE, callTypeLog2R = "summary",
  callThreshLog2R = c(-0.15, 0.15), callLegend = FALSE,
  callLegendPosition = "bottom", callLegendHoriz = TRUE, pointShape = ".",
  pointSize = 1, lossColor = "orange2", gainColor = "blue",
  pointColor = "black", neutralColor = "white", missingColor = "grey20",
  xlab = ifelse(is.null(geneInfo), "Chromosome", "Gene"), xaxt = "s",
  xaxSize = 7, ylab = NULL, yaxSize = 10, ylim = NULL, yaxp = NULL,
  main = NULL, mainSize = 10, ...)
}
\arguments{
\item{object}{CNAclinicData object}
}
\value{
a list containing ggplot objects (1 per sample, list named by sampleNames)
}
\description{
plotSampleData XXXXX
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
