% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run up to 4 different copy number segmentation algorithms.}
\usage{
runSegmentation(x, isLogTransformed = TRUE, genome = NULL,
  segmentType = c("CBS", "LACBS", "HMM", "PLS"), summaryMethod = "mean",
  segmentsToSummarise = segmentType, callTypeLog2R = "summary",
  callThreshLog2R = c(-0.15, 0.15), minimumBinsPerSegment = 3,
  gapLength = 1e+06, normalizeSegmentedBins = TRUE, inter = c(-0.1, 0.1),
  alpha = 1e-10, undo.splits = "sdundo", undo.SD = 1,
  segmentStatistic = "seg.mean")
}
\arguments{
\item{x}{Either a \code{\link{CNAclinicData}} object from the output of
\code{\link{processForSegmentation}} function or a \code{data.frame} with
specific columns (see details) or a \code{\link{QDNAseqCopyNumbers}} object.}

\item{isLogTransformed}{if TRUE (default) data is assumed to be log2
transformed. If FALSE, transformation is carried out prior to segmentation.}

\item{genome}{Genome build used to align sequencing reads.}

\item{segmentType}{One or more algorithms for segmentation.
i.e. \code{c("CBS", "HMM", "PLS", "LACBS")}}

\item{summaryMethod}{Summarization method for ensemble segmentation.
i.e. One of c("mean", "median", "min", "max", "Q1", "Q3")}

\item{segmentsToSummarise}{X}

\item{callTypeLog2R}{Segment type used to call CNAs (default "summary")}

\item{callThreshLog2R}{Thresholds used to call segments as a "loss" or "gain".
Defaults to c(-0.15, 0.15)}

\item{minimumBinsPerSegment}{Minimum number of bins in each segment,
default is 3. Argument is specific to PLS algorithm.}

\item{gapLength}{Minimum length in base-pairs between the two closest loci
to consider a region to be a "gap". (default 1000000)
Argument is specific to LACBS algorithm.}

\item{normalizeSegmentedBins}{Normalizes the segmented bins if TRUE (default).}

\item{inter}{The interval in which the function should search for the
normal level. Utilised in \code{normalizeSegmentedBins}.}

\item{alpha}{Significance levels for the test to accept change-points.
Default is 1e-10}

\item{undo.splits}{A character string specifying how change-points are to be
undone, if at all. Default is "sdundo", which undoes splits that are not at
least this many SDs apart. Other choices are "prune", which uses a sum of
squares criterion, and "none".}

\item{undo.SD}{The number of SDs between means to keep a split if
undo.splits="sdundo". Default is 1.0.}

\item{segmentStatistic}{Default is "seg.mean".}
}
\value{
Returns an object of class \code{\link{CNAclinicData}} with
segmentation results from the chosen algorithms as well as the
summarised result.
}
\description{
\code{runSegmentation} Segments binned and normalized
copy number values using multiple algorithms and allows the summarization
of the segmentation results. Automatically calculates the "gain", "loss" calls
for each segment.
}
\details{
If \code{x} is a data.frame it should contain the columns:
\code{chromosome}, \code{start}, \code{end} and the optional column
\code{usebin} followed by separate columns named after each different sample
contaning normalized and log-transformed copy number measurements.
\code{chromosome} should be a character vector containing only 1-22, X, Y
while all other columns should contain numerical values.
The genomic coordinates should belong to fixed-width bins.
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
\author{
Dineika Chandrananda
}
