% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsCNA.R
\docType{methods}
\name{statsCNA,CNAclinicData-method}
\alias{statsCNA,CNAclinicData-method}
\title{Quantify and quality check copy-number profiles.}
\usage{
\S4method{statsCNA}{CNAclinicData}(object, measure = c("FGA", "MAPD"),
  digits = 3)
}
\arguments{
\item{object}{a CNAclinicData object containing copy number calls}

\item{measure}{Either "FGA", "MAPD" or both}
}
\value{
a matrix containing the values for the chosen measure(s) for each
sample
}
\description{
Calculate the fraction of genome altered (FGA) and/or the
median of the absolute values of all pairwise differences (MAPD) between
adjacent genomic bin values.
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
