% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\docType{methods}
\name{subsetData,CNAclinicData-method}
\alias{subsetData,CNAclinicData-method}
\title{Extract data belonging to a subset of samples in a CNAclinic object}
\usage{
\S4method{subsetData}{CNAclinicData}(object,
  sampleNames = CNAclinic::sampleNames(object))
}
\arguments{
\item{object}{a \code{\link{CNAclinicData}} object}

\item{sampleNames}{a subset of sample names in \code{x}}
}
\value{
Returns an object of class \code{\link{CNAclinicData}} with data
belonging to the \code{sampleNames}
}
\description{
This function allows splitting up a CNAclinic object into different sample
subsets.
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }

}
