% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summSegments.R
\docType{methods}
\name{summSegments,CNAclinicData-method}
\alias{summSegments,CNAclinicData-method}
\title{Summarise segments from multiple algorithms at every genomic bin}
\usage{
\S4method{summSegments}{CNAclinicData}(object, segmentsToSummarise = c("CBS",
  "LACBS", "HMM", "PLS"), summaryMethod = "mean")
}
\arguments{
\item{object}{CNAclinicData.}
}
\value{
An updated CNAclinicData object.
}
\description{
Summarise segments from multiple algorithms at every genomic bin
}
\examples{
     \dontrun{
      vignette("CNAclinic")
     }


}
